backend/requirements.txt
fastapi==0.78.0
Pillow==9.1.1
uvicorn==0.17.6
python-multipart
ninja==1.10.2.3
torch==1.11.0
torchvision
dlib
scikit-learn==0.22
scipy==1.4.1
lpips==0.1.4
opencv-python-headless==4.5.5.62
wandb==0.12.9
numpy==1.19.5
matplotlib==3.2.2